+++ pyteal/ast/__init__.py
@@ -258,3 +258,10 @@ __all__ = [
     "Break",
     "Continue",
 ]
+
+# abi types
+from .abi_type import ABIType
+from .abi_collections import Tuple, FixedArray, DynamicArray
+from .abi_bytes import String, Address
+from .abi_uint import Uint16, Uint32, Uint64, Uint128, Uint256, Uint512
+from .abi_ufixed import UFixed
